+++ dif-data-classes/src/commonMain/kotlin/at/asitplus/dif/ClaimFormat.kt
@@ -1,5 +1,6 @@
 package at.asitplus.dif
 
+import at.asitplus.dif.ClaimFormat.entries
 import kotlinx.serialization.Serializable
 
 /**
@@ -12,7 +13,9 @@ enum class ClaimFormat(val text: String) {
     JWT("jwt"),
     JWT_VC("jwt_vc"),
     JWT_VP("jwt_vp"),
+    @Deprecated("Deprecated in SD-JWT VC since draft 06", replaceWith = ReplaceWith("SD_JWT"))
     JWT_SD("vc+sd-jwt"),
+    SD_JWT("dc+sd-jwt"),
     LDP("ldp"),
     LDP_VC("ldp_vc"),
     LDP_VP("ldp_vp"),