+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/OpenId4VpVerifier.kt
@@ -97,6 +97,7 @@ class OpenId4VpVerifier(
                 msoMdoc = containerJwt,
                 jwtVp = containerJwt,
                 jwtSd = containerSdJwt,
+                sdJwt = containerSdJwt
             )
         )
     }
@@ -309,7 +310,7 @@ class OpenId4VpVerifier(
 
     private fun RequestOptionsCredential.toFormatHolder() = when (representation) {
         ConstantIndex.CredentialRepresentation.PLAIN_JWT -> FormatHolder(jwtVp = containerJwt)
-        ConstantIndex.CredentialRepresentation.SD_JWT -> FormatHolder(jwtSd = containerSdJwt)
+        ConstantIndex.CredentialRepresentation.SD_JWT -> FormatHolder(jwtSd = containerSdJwt, sdJwt = containerSdJwt)
         ConstantIndex.CredentialRepresentation.ISO_MDOC -> FormatHolder(msoMdoc = containerJwt)
     }
 
@@ -503,6 +504,7 @@ class OpenId4VpVerifier(
      * [OpenID for VCI](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html),
      * as referenced by [OpenID for VP](https://openid.net/specs/openid-4-verifiable-presentations-1_0.html).
      */
+    @Suppress("DEPRECATION")
     private suspend fun verifyPresentationResult(
         descriptor: PresentationSubmissionDescriptor,
         relatedPresentation: JsonElement,
@@ -511,7 +513,7 @@ class OpenId4VpVerifier(
         clientId: String?,
         responseUrl: String?,
     ) = when (descriptor.format) {
-        ClaimFormat.JWT_SD -> verifier.verifyPresentationSdJwt(
+        ClaimFormat.JWT_SD, ClaimFormat.SD_JWT -> verifier.verifyPresentationSdJwt(
             input = SdJwtSigned.Companion.parse(relatedPresentation.jsonPrimitive.content)
                 ?: throw IllegalArgumentException("relatedPresentation"),
             challenge = expectedNonce