+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/PresentationFactory.kt
@@ -238,11 +238,22 @@ internal class PresentationFactory(
             }
         }
 
-    private fun FormatHolder.isMissingFormatSupport(claimFormat: ClaimFormat) =
-        when (claimFormat) {
+    @Suppress("DEPRECATION")
+    private fun FormatHolder.isMissingFormatSupport(claimFormat: ClaimFormat): Boolean {
+        return when (claimFormat) {
             ClaimFormat.JWT_VP -> jwtVp?.algorithms?.let { !it.contains(jwsService.algorithm) } ?: false
-            ClaimFormat.JWT_SD -> jwtSd?.algorithms?.let { !it.contains(jwsService.algorithm) } ?: false
+            ClaimFormat.JWT_SD, ClaimFormat.SD_JWT -> {
+                if (jwtSd?.algorithms?.contains(jwsService.algorithm) == false) return true
+                if (jwtSd?.sdJwtAlgorithms?.contains(jwsService.algorithm) == false) return true
+                if (jwtSd?.kbJwtAlgorithms?.contains(jwsService.algorithm) == false) return true
+                if (sdJwt?.algorithms?.contains(jwsService.algorithm) == false) return true
+                if (sdJwt?.sdJwtAlgorithms?.contains(jwsService.algorithm) == false) return true
+                if (sdJwt?.kbJwtAlgorithms?.contains(jwsService.algorithm) == false) return true
+                return false
+            }
+
             ClaimFormat.MSO_MDOC -> msoMdoc?.algorithms?.let { !it.contains(jwsService.algorithm) } ?: false
             else -> false
         }
+    }
 }