+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciInteropTest.kt
@@ -202,6 +202,7 @@ class OidvciInteropTest : FunSpec({
             .entries.first { it.key == credentialOffer.configurationIds.first() }.toPair()
 
         val credential = credentialConfig.second
+        @Suppress("DEPRECATION")
         credential.format shouldBe CredentialFormatEnum.VC_SD_JWT
         credential.scope shouldBe "eu.europa.ec.eudi.pid_vc_sd_jwt"
         credential.supportedBindingMethods!!.shouldHaveSingleElement("jwk")