+++ CHANGELOG.md
@@ -50,6 +50,8 @@ Release 5.3.0:
   - Move `RequestObjectJwsVerifier` from `at.asitplus.wallet.lib.oidc` to `at.asitplus.wallet.lib.openid`
   - Move `RemoteResourceRetrieverFunction` from `at.asitplus.wallet.lib.oidc` to `at.asitplus.wallet.lib`
   - Move `AuthorizationResponsePreparationState` from `at.asitplus.wallet.lib.oidc.helpers` to `at.asitplus.wallet.lib.openid`
+- OpenID4VP draft 23:
+  - Support credential format identifier `dc+sd-jwt` in addition to `vc+sd-jwt`
 - General cleanup:
   - Remove `SchemaIndex`
   - Update signum to 3.12.1